+++ integration_test/support/models.exs
@@ -11,8 +11,8 @@ defmodule Ecto.Integration.Model do
     end
   end
 
-  def callback_check(changeset) do
-    Process.put(:callback_check, true)
+  def pdict_store(changeset, key, val) do
+    Process.put(key, val)
     changeset
   end
 end
@@ -42,8 +42,8 @@ defmodule Ecto.Integration.Post do
     field :bid, :binary_id
     field :uuid, Ecto.UUID, autogenerate: true
     field :meta, :map
-    has_many :comments, Ecto.Integration.Comment, dependent: :delete_all
-    has_one :permalink, Ecto.Integration.Permalink, dependent: :fetch_and_delete
+    has_many :comments, Ecto.Integration.Comment, on_delete: :delete_all
+    has_one :permalink, Ecto.Integration.Permalink, on_delete: :fetch_and_delete
     has_many :comments_authors, through: [:comments, :author]
     belongs_to :author, Ecto.Integration.User
     timestamps
@@ -86,7 +86,7 @@ defmodule Ecto.Integration.Comment do
   end
 
   optimistic_lock :lock_version
-  before_delete Ecto.Integration.Model, :callback_check, []
+  before_delete Ecto.Integration.Model, :pdict_store, [__MODULE__, :on_delete]
 end
 
 defmodule Ecto.Integration.Permalink do
@@ -105,7 +105,7 @@ defmodule Ecto.Integration.Permalink do
     has_many :post_comments_authors, through: [:post, :comments_authors]
   end
 
-  before_delete Ecto.Integration.Model, :callback_check, []
+  before_delete Ecto.Integration.Model, :pdict_store, [__MODULE__, :on_delete]
 end
 
 defmodule Ecto.Integration.User do
@@ -121,8 +121,8 @@ defmodule Ecto.Integration.User do
 
   schema "users" do
     field :name, :string
-    has_many :comments, Ecto.Integration.Comment, foreign_key: :author_id, dependent: :nilify_all
-    has_many :posts, Ecto.Integration.Post, dependent: :nothing, foreign_key: :author_id
+    has_many :comments, Ecto.Integration.Comment, foreign_key: :author_id, on_delete: :nilify_all
+    has_many :posts, Ecto.Integration.Post, on_delete: :nothing, foreign_key: :author_id
     belongs_to :custom, Ecto.Integration.Custom, references: :bid, type: :binary_id
     timestamps
   end